refactor: move client integration tests to simapp
